requirements.txt
asgiref==3.8.1
dj-database-url==3.0.0
Django==5.0.7
django-cors-headers==4.4.0
django-jazzmin==3.0.0
djangorestframework==3.15.2
gunicorn==23.0.0
packaging==25.0
pillow==10.4.0
psycopg2-binary==2.9.10
sqlparse==0.5.1
typing_extensions==4.14.0
